add native base and react navigation
