+++ LICENSE
@@ -1,8 +1,8 @@
 MIT License
 
 Copyright (C) 2005-2017 Anthony Minessale II
-Copyright (C) 2018 Seven Du <dujinfang@x-y-t.cn(opens in new tab)>
-Copyright (C) 2018 Xueyun Jiang <jiangxueyun@x-y-t.cn(opens in new tab)>
+Copyright (C) 2018 Seven Du <dujinfang@x-y-t.cn>
+Copyright (C) 2018 Xueyun Jiang <jiangxueyun@x-y-t.cn>
 Copyright (C) 2019 Telnyx, Inc. (https://www.telnyx.com)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
