+++ examples/react-native/index.js
@@ -1,6 +1,7 @@
 /**
  * @format
  */
+import 'react-native-gesture-handler';
 import 'react-native-get-random-values';
 import {AppRegistry} from 'react-native';
 import App from './App';
