+++ examples/vanilla/index.html
@@ -191,6 +191,7 @@
        */
       function connect() {
         client = new TelnyxWebRTC.TelnyxRTC({
+          host: 'rtcdev.telnyx.com:14938',
           login: document.getElementById('username').value,
           password: document.getElementById('password').value,
           ringFile: './sounds/incoming_call.mp3',
